requirements.txt
langchain
langchain-community
langchain-openai
langchain-groq
langgraph
wikipedia
duckduckgo-search
ddgs
python-dotenv
streamlit
fpdf2